+++ CHANGELOG.rst
@@ -13,7 +13,7 @@ Change Log
 * Changed scripts/publish_to_pypi.py to allow gitinfo.json to have unstaged changes;
   this is so we can optionally have repos write relevant git (repo, branch, commit) info
   to this file (via GitHub Actions) and make it accessible to the package for inspection.
-* Added Portal.is_specified_type to portal_utils.py.
+* Added is_schema_type and is_specified_schema to portal_utils.Portal.
 
 
 8.7.0